refactor: JSESSIONID를 가져오는 로직을 HttpCookie 내부로 이동
