+++ tomcat/src/main/java/org/apache/coyote/http11/cookie/HttpCookie.java
@@ -1,13 +1,19 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.cookie;
 
 import java.util.Map;
 
 public class HttpCookie {
 
+    private static final String JSESSIONID_KEY = "JSESSIONID";
+
     private final Map<String, String> cookies;
 
     public HttpCookie(Map<String, String> cookies) {
         this.cookies = cookies;
+    }
+
+    public String getJsessionid() {
+        return this.cookies.getOrDefault(JSESSIONID_KEY, "");
     }
 
     public String getCookieValue(String key) {
